36 fields determine sharednt handler initialisation (#42)
The main purpose of this PR is change SharedNT to add Rules to the CompositeHandler based on information extracted from information extracted from the Rules through introspection. This replaces lots of if statements and turns it into a general approach which checks if a Rule applies to certain NT Types (incomplete), fields, or must be manually specified through a constructor.

In more detail:
* Creates new files in `p4pillon/nt` which holds information about the specification of NT Types (`specs`) and has tools for identifying NT Types (`identify`)
* Removes obsolete code from `nthandlers`
* Adds new member variables to Rules base and concrete classes. These are documented in the `BaseRule` class.
* Updates tests to reflect these changes. In addition, integration tests are renamed to prevent name clashing with unit tests. `tests/integration/test_mixed_concurrency.py` is updated to timeout so that errors cause it to crash rather than hang indefinitely.